refactor: move batch item struct types to a new namespace
