+++ Testing/VelaptorTests/Services/LineBatchingManagerTests.cs
@@ -18,7 +18,7 @@ using Moq;
 using Velaptor;
 using Velaptor.Exceptions;
 using Velaptor.Factories;
-using Velaptor.OpenGL;
+using Velaptor.OpenGL.Batching;
 using Velaptor.ReactableData;
 using Velaptor.Services;
 using Xunit;
