+++ Velaptor/Graphics/Renderers/RectangleRenderer.cs
@@ -11,6 +11,7 @@ using Factories;
 using Guards;
 using NativeInterop.OpenGL;
 using OpenGL;
+using OpenGL.Batching;
 using OpenGL.Buffers;
 using OpenGL.Shaders;
 using Services;
