+++ Velaptor/Graphics/Renderers/TextureRenderer.cs
@@ -13,6 +13,7 @@ using Factories;
 using Guards;
 using NativeInterop.OpenGL;
 using OpenGL;
+using OpenGL.Batching;
 using OpenGL.Buffers;
 using OpenGL.Shaders;
 using Services;
