+++ Velaptor/OpenGL/Batching/FontGlyphBatchItem.cs
@@ -1,8 +1,8 @@
-// <copyright file="TextureBatchItem.cs" company="KinsonDigital">
+﻿// <copyright file="FontGlyphBatchItem.cs" company="KinsonDigital">
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.OpenGL;
+namespace Velaptor.OpenGL.Batching;
 
 using System;
 using System.Diagnostics.CodeAnalysis;
@@ -12,25 +12,26 @@ using System.Text;
 using Graphics;
 
 /// <summary>
-/// A single texture batch item that can be rendered to the screen.
+/// A single item in a batch of glyph items that can be rendered to the screen.
 /// </summary>
-internal readonly struct TextureBatchItem : IEquatable<TextureBatchItem>
+internal readonly struct FontGlyphBatchItem : IEquatable<FontGlyphBatchItem>
 {
     /// <summary>
-    /// Initializes a new instance of the <see cref="TextureBatchItem"/> struct.
+    /// Initializes a new instance of the <see cref="FontGlyphBatchItem"/> struct.
     /// </summary>
     /// <param name="srcRect">The rectangular section inside of a texture to render.</param>
-    /// <param name="destRect">The destination rectangular area of where to render the texture on the screen.</param>
-    /// <param name="size">The size of the rendered texture.</param>
-    /// <param name="angle">The angle of the texture in degrees.</param>
-    /// <param name="tintColor">The color to apply to the entire texture.</param>
-    /// <param name="effects">The type of effects to apply to a texture.</param>
-    /// <param name="textureId">The ID of the texture.</param>
-    /// <param name="layer">The layer where a texture will be rendered.</param>
-    public TextureBatchItem(
-        in
+    /// <param name="destRect">The destination rectangular area of where to render the glyph on the screen.</param>
+    /// <param name="glyph">The font glyph.</param>
+    /// <param name="size">The size of the glyph texture to be rendered.</param>
+    /// <param name="angle">The angle in degrees of the glyph texture.</param>
+    /// <param name="tintColor">The color to apply to the entire glyph texture.</param>
+    /// <param name="effects">The type of effects to apply to the glyph texture when rendering.</param>
+    /// <param name="textureId">The ID of the font atlas texture.</param>
+    /// <param name="layer">The layer where the shape will be rendered.</param>
+    public FontGlyphBatchItem(
         RectangleF srcRect,
         RectangleF destRect,
+        char glyph,
         float size,
         float angle,
         Color tintColor,
@@ -40,6 +41,7 @@ internal readonly struct TextureBatchItem : IEquatable<TextureBatchItem>
     {
         SrcRect = srcRect;
         DestRect = destRect;
+        Glyph = glyph;
         Size = size;
         Angle = angle;
         TintColor = tintColor;
@@ -54,37 +56,44 @@ internal readonly struct TextureBatchItem : IEquatable<TextureBatchItem>
     public RectangleF SrcRect { get; }
 
     /// <summary>
-    /// Gets the destination rectangular area of where to render the texture on the screen.
+    /// Gets the destination rectangular area of where to render the glyph on the screen.
     /// </summary>
     public RectangleF DestRect { get; }
 
     /// <summary>
-    /// Gets the size of the rendered texture.
+    /// Gets the font glyph.
     /// </summary>
+    public char Glyph { get; }
+
+    /// <summary>
+    /// Gets the size of the glyph texture to be rendered.
+    /// </summary>
+    /// <remarks>This must be a value between 0 and 1.</remarks>
     public float Size { get; }
 
     /// <summary>
-    /// Gets the angle of the texture in degrees.
+    /// Gets the angle in degrees of the glyph texture.
     /// </summary>
+    /// <remarks>Needs to be a value between 0 and 360.</remarks>
     public float Angle { get; }
 
     /// <summary>
-    /// Gets the color to apply to an entire texture.
+    /// Gets the color to apply to the entire glyph texture.
     /// </summary>
     public Color TintColor { get; }
 
     /// <summary>
-    /// Gets the type of effects to apply to a texture.
+    /// Gets the type of effects to apply to the glyph texture when rendering.
     /// </summary>
-    public RenderEffects Effects { get; } = RenderEffects.None;
+    public RenderEffects Effects { get; }
 
     /// <summary>
-    /// Gets the ID of the texture.
+    /// Gets the ID of the font atlas texture.
     /// </summary>
     public uint TextureId { get; }
 
     /// <summary>
-    /// Gets the layer that the a texture will be rendered on.
+    /// Gets the layer where the shape will be rendered.
     /// </summary>
     /// <remarks>
     ///     <para>
@@ -122,7 +131,7 @@ internal readonly struct TextureBatchItem : IEquatable<TextureBatchItem>
     /// <param name="left">The left operand compared with the right operand.</param>
     /// <param name="right">The right operand compared with the left operand.</param>
     /// <returns>True if both operands are equal.</returns>
-    public static bool operator ==(TextureBatchItem left, TextureBatchItem right) => left.Equals(right);
+    public static bool operator ==(FontGlyphBatchItem left, FontGlyphBatchItem right) => left.Equals(right);
 
     /// <summary>
     /// Returns a value indicating whether or not the <paramref name="left"/> operand is not equal to the <paramref name="right"/> operand.
@@ -130,24 +139,25 @@ internal readonly struct TextureBatchItem : IEquatable<TextureBatchItem>
     /// <param name="left">The left operand compared with the right operand.</param>
     /// <param name="right">The right operand compared with the left operand.</param>
     /// <returns>True if both operands are not equal.</returns>
-    public static bool operator !=(TextureBatchItem left, TextureBatchItem right) => !(left == right);
+    public static bool operator !=(FontGlyphBatchItem left, FontGlyphBatchItem right) => !(left == right);
 
     /// <summary>
-    /// Gets a value indicating whether or not the current <see cref="TextureBatchItem"/> is empty.
+    /// Gets a value indicating whether or not the current <see cref="FontGlyphBatchItem"/> is empty.
     /// </summary>
     /// <returns>True if empty.</returns>
     public bool IsEmpty() =>
         TextureId == 0 &&
         Size == 0f &&
         Angle == 0f &&
-        Effects is 0 or RenderEffects.None &&
         Layer == 0 &&
+        Effects is 0 or RenderEffects.None &&
+        Glyph == '\0' &&
         SrcRect.IsEmpty &&
         DestRect.IsEmpty &&
         TintColor.IsEmpty;
 
     /// <inheritdoc cref="IEquatable{T}.Equals(T?)"/>
-    public bool Equals(TextureBatchItem other) =>
+    public bool Equals(FontGlyphBatchItem other) =>
         SrcRect.Equals(other.SrcRect) &&
         DestRect.Equals(other.DestRect) &&
         Size.Equals(other.Size) &&
@@ -155,23 +165,25 @@ internal readonly struct TextureBatchItem : IEquatable<TextureBatchItem>
         TintColor.Equals(other.TintColor) &&
         Effects == other.Effects &&
         TextureId == other.TextureId &&
+        Glyph == other.Glyph &&
         Layer == other.Layer;
 
     /// <inheritdoc cref="object.Equals(object?)"/>
-    public override bool Equals(object? obj) => obj is TextureBatchItem other && Equals(other);
+    public override bool Equals(object? obj) => obj is FontGlyphBatchItem other && Equals(other);
 
     /// <inheritdoc cref="object.GetHashCode"/>
     [ExcludeFromCodeCoverage(Justification = "Cannot test because hash codes do not return repeatable results.")]
     public override int GetHashCode()
         => HashCode.Combine(
             HashCode.Combine(
-            SrcRect,
-            DestRect,
-            Size,
-            Angle,
-            TintColor,
-            (int)Effects,
-            TextureId),
+                SrcRect,
+                DestRect,
+                Size,
+                Angle,
+                TintColor,
+                (int)Effects,
+                TextureId),
+            Glyph,
             Layer);
 
     /// <inheritdoc/>
@@ -179,7 +191,7 @@ internal readonly struct TextureBatchItem : IEquatable<TextureBatchItem>
     {
         var result = new StringBuilder();
 
-        result.AppendLine("Texture Batch Item Values:");
+        result.AppendLine("Font Batch Item Values:");
         result.AppendLine($"Src Rect: {SrcRect.ToString()}");
         result.AppendLine($"Dest Rect: {DestRect.ToString()}");
         result.AppendLine($"Size: {Size.ToString(CultureInfo.InvariantCulture)}");
@@ -188,6 +200,7 @@ internal readonly struct TextureBatchItem : IEquatable<TextureBatchItem>
             $"Tint Clr: {{A={TintColor.A},R={TintColor.R},G={TintColor.G},B={TintColor.B}}}");
         result.AppendLine($"Effects: {Effects.ToString()}");
         result.AppendLine($"Texture ID: {TextureId.ToString()}");
+        result.AppendLine($"Glyph: {Glyph}");
         result.Append($"Layer: {Layer}");
 
         return result.ToString();
