+++ Velaptor/OpenGL/Batching/TextureBatchItem.cs
@@ -1,8 +1,8 @@
-﻿// <copyright file="FontGlyphBatchItem.cs" company="KinsonDigital">
+// <copyright file="TextureBatchItem.cs" company="KinsonDigital">
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.OpenGL;
+namespace Velaptor.OpenGL.Batching;
 
 using System;
 using System.Diagnostics.CodeAnalysis;
@@ -12,26 +12,25 @@ using System.Text;
 using Graphics;
 
 /// <summary>
-/// A single item in a batch of glyph items that can be rendered to the screen.
+/// A single texture batch item that can be rendered to the screen.
 /// </summary>
-internal readonly struct FontGlyphBatchItem : IEquatable<FontGlyphBatchItem>
+internal readonly struct TextureBatchItem : IEquatable<TextureBatchItem>
 {
     /// <summary>
-    /// Initializes a new instance of the <see cref="FontGlyphBatchItem"/> struct.
+    /// Initializes a new instance of the <see cref="TextureBatchItem"/> struct.
     /// </summary>
     /// <param name="srcRect">The rectangular section inside of a texture to render.</param>
-    /// <param name="destRect">The destination rectangular area of where to render the glyph on the screen.</param>
-    /// <param name="glyph">The font glyph.</param>
-    /// <param name="size">The size of the glyph texture to be rendered.</param>
-    /// <param name="angle">The angle in degrees of the glyph texture.</param>
-    /// <param name="tintColor">The color to apply to the entire glyph texture.</param>
-    /// <param name="effects">The type of effects to apply to the glyph texture when rendering.</param>
-    /// <param name="textureId">The ID of the font atlas texture.</param>
-    /// <param name="layer">The layer where the shape will be rendered.</param>
-    public FontGlyphBatchItem(
+    /// <param name="destRect">The destination rectangular area of where to render the texture on the screen.</param>
+    /// <param name="size">The size of the rendered texture.</param>
+    /// <param name="angle">The angle of the texture in degrees.</param>
+    /// <param name="tintColor">The color to apply to the entire texture.</param>
+    /// <param name="effects">The type of effects to apply to a texture.</param>
+    /// <param name="textureId">The ID of the texture.</param>
+    /// <param name="layer">The layer where a texture will be rendered.</param>
+    public TextureBatchItem(
+        in
         RectangleF srcRect,
         RectangleF destRect,
-        char glyph,
         float size,
         float angle,
         Color tintColor,
@@ -41,7 +40,6 @@ internal readonly struct FontGlyphBatchItem : IEquatable<FontGlyphBatchItem>
     {
         SrcRect = srcRect;
         DestRect = destRect;
-        Glyph = glyph;
         Size = size;
         Angle = angle;
         TintColor = tintColor;
@@ -56,44 +54,37 @@ internal readonly struct FontGlyphBatchItem : IEquatable<FontGlyphBatchItem>
     public RectangleF SrcRect { get; }
 
     /// <summary>
-    /// Gets the destination rectangular area of where to render the glyph on the screen.
+    /// Gets the destination rectangular area of where to render the texture on the screen.
     /// </summary>
     public RectangleF DestRect { get; }
 
     /// <summary>
-    /// Gets the font glyph.
+    /// Gets the size of the rendered texture.
     /// </summary>
-    public char Glyph { get; }
-
-    /// <summary>
-    /// Gets the size of the glyph texture to be rendered.
-    /// </summary>
-    /// <remarks>This must be a value between 0 and 1.</remarks>
     public float Size { get; }
 
     /// <summary>
-    /// Gets the angle in degrees of the glyph texture.
+    /// Gets the angle of the texture in degrees.
     /// </summary>
-    /// <remarks>Needs to be a value between 0 and 360.</remarks>
     public float Angle { get; }
 
     /// <summary>
-    /// Gets the color to apply to the entire glyph texture.
+    /// Gets the color to apply to an entire texture.
     /// </summary>
     public Color TintColor { get; }
 
     /// <summary>
-    /// Gets the type of effects to apply to the glyph texture when rendering.
+    /// Gets the type of effects to apply to a texture.
     /// </summary>
-    public RenderEffects Effects { get; }
+    public RenderEffects Effects { get; } = RenderEffects.None;
 
     /// <summary>
-    /// Gets the ID of the font atlas texture.
+    /// Gets the ID of the texture.
     /// </summary>
     public uint TextureId { get; }
 
     /// <summary>
-    /// Gets the layer where the shape will be rendered.
+    /// Gets the layer that the a texture will be rendered on.
     /// </summary>
     /// <remarks>
     ///     <para>
@@ -131,7 +122,7 @@ internal readonly struct FontGlyphBatchItem : IEquatable<FontGlyphBatchItem>
     /// <param name="left">The left operand compared with the right operand.</param>
     /// <param name="right">The right operand compared with the left operand.</param>
     /// <returns>True if both operands are equal.</returns>
-    public static bool operator ==(FontGlyphBatchItem left, FontGlyphBatchItem right) => left.Equals(right);
+    public static bool operator ==(TextureBatchItem left, TextureBatchItem right) => left.Equals(right);
 
     /// <summary>
     /// Returns a value indicating whether or not the <paramref name="left"/> operand is not equal to the <paramref name="right"/> operand.
@@ -139,25 +130,24 @@ internal readonly struct FontGlyphBatchItem : IEquatable<FontGlyphBatchItem>
     /// <param name="left">The left operand compared with the right operand.</param>
     /// <param name="right">The right operand compared with the left operand.</param>
     /// <returns>True if both operands are not equal.</returns>
-    public static bool operator !=(FontGlyphBatchItem left, FontGlyphBatchItem right) => !(left == right);
+    public static bool operator !=(TextureBatchItem left, TextureBatchItem right) => !(left == right);
 
     /// <summary>
-    /// Gets a value indicating whether or not the current <see cref="FontGlyphBatchItem"/> is empty.
+    /// Gets a value indicating whether or not the current <see cref="TextureBatchItem"/> is empty.
     /// </summary>
     /// <returns>True if empty.</returns>
     public bool IsEmpty() =>
         TextureId == 0 &&
         Size == 0f &&
         Angle == 0f &&
-        Layer == 0 &&
         Effects is 0 or RenderEffects.None &&
-        Glyph == '\0' &&
+        Layer == 0 &&
         SrcRect.IsEmpty &&
         DestRect.IsEmpty &&
         TintColor.IsEmpty;
 
     /// <inheritdoc cref="IEquatable{T}.Equals(T?)"/>
-    public bool Equals(FontGlyphBatchItem other) =>
+    public bool Equals(TextureBatchItem other) =>
         SrcRect.Equals(other.SrcRect) &&
         DestRect.Equals(other.DestRect) &&
         Size.Equals(other.Size) &&
@@ -165,25 +155,23 @@ internal readonly struct FontGlyphBatchItem : IEquatable<FontGlyphBatchItem>
         TintColor.Equals(other.TintColor) &&
         Effects == other.Effects &&
         TextureId == other.TextureId &&
-        Glyph == other.Glyph &&
         Layer == other.Layer;
 
     /// <inheritdoc cref="object.Equals(object?)"/>
-    public override bool Equals(object? obj) => obj is FontGlyphBatchItem other && Equals(other);
+    public override bool Equals(object? obj) => obj is TextureBatchItem other && Equals(other);
 
     /// <inheritdoc cref="object.GetHashCode"/>
     [ExcludeFromCodeCoverage(Justification = "Cannot test because hash codes do not return repeatable results.")]
     public override int GetHashCode()
         => HashCode.Combine(
             HashCode.Combine(
-                SrcRect,
-                DestRect,
-                Size,
-                Angle,
-                TintColor,
-                (int)Effects,
-                TextureId),
-            Glyph,
+            SrcRect,
+            DestRect,
+            Size,
+            Angle,
+            TintColor,
+            (int)Effects,
+            TextureId),
             Layer);
 
     /// <inheritdoc/>
@@ -191,7 +179,7 @@ internal readonly struct FontGlyphBatchItem : IEquatable<FontGlyphBatchItem>
     {
         var result = new StringBuilder();
 
-        result.AppendLine("Font Batch Item Values:");
+        result.AppendLine("Texture Batch Item Values:");
         result.AppendLine($"Src Rect: {SrcRect.ToString()}");
         result.AppendLine($"Dest Rect: {DestRect.ToString()}");
         result.AppendLine($"Size: {Size.ToString(CultureInfo.InvariantCulture)}");
@@ -200,7 +188,6 @@ internal readonly struct FontGlyphBatchItem : IEquatable<FontGlyphBatchItem>
             $"Tint Clr: {{A={TintColor.A},R={TintColor.R},G={TintColor.G},B={TintColor.B}}}");
         result.AppendLine($"Effects: {Effects.ToString()}");
         result.AppendLine($"Texture ID: {TextureId.ToString()}");
-        result.AppendLine($"Glyph: {Glyph}");
         result.Append($"Layer: {Layer}");
 
         return result.ToString();
