+++ Velaptor/OpenGL/Buffers/TextureGPUBuffer.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using System.Numerics;
+using Batching;
 using Carbonate.UniDirectional;
 using Exceptions;
 using Factories;
