+++ Velaptor/Services/TextureBatchingManager.cs
@@ -14,7 +14,7 @@ using Content;
 using Exceptions;
 using Factories;
 using Guards;
-using OpenGL;
+using OpenGL.Batching;
 using ReactableData;
 
 /// <summary>
